assistant_service/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic>=2.11,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.28,<1.0
